+++ src/pyws/errors.py
@@ -1,55 +1,70 @@
+from pyws.utils import DefaultStrImplemntationMixin
+
 ET_CLIENT = 1
 ET_SERVER = 2
 
 
-class ConfigurationError(Exception):
-    pass
+class Error(DefaultStrImplemntationMixin, Exception):
 
-
-class ServerAlreadyRegistered(ConfigurationError):
-    def __str__(self):
-        return 'Server %s has already been registered' % self.args[0]
-
-
-class DefaultServerAlreadyRegistered(ConfigurationError):
-    def __str__(self):
-        return 'Default server has already been registered'
-
-
-class SettingNotDefined(ConfigurationError):
-    def __str__(self):
-        return 'Setting %s is not defined' % self.args[0]
-
-
-class NoProtocolsRegistered(ConfigurationError):
-    def __str__(self):
-        return 'Server has no protocols'
-
-
-class BadProtocol(ConfigurationError):
-
-    def __str__(self):
-        return 'Bad protocol: %s' % self.args[0]
-
-
-class FunctionAlreadyRegistered(ConfigurationError):
-
-    def __str__(self):
-        return 'Function %s is already registered' % self.args[0]
-
-
-class BadFunction(ConfigurationError):
-
-    def __str__(self):
-        return 'Bad function: %s' % self.args[0]
-
-
-class Error(Exception):
     error_type = ET_SERVER
+
+    def __unicode__(self):
+        if self.__doc__:
+            return unicode(self.__doc__.strip() % self.args)
+        if self.args:
+            return self.args[0]
+        return 'unknown error'
 
 
 class ClientErrorTypeMixin(Exception):
     error_type = ET_CLIENT
+
+
+class ConfigurationError(Error):
+    def __unicode__(self):
+        return unicode(self.args[0])
+
+
+class ServerAlreadyRegistered(ConfigurationError):
+    """
+    Server '%s' has already been registered
+    """
+
+
+class DefaultServerAlreadyRegistered(ConfigurationError):
+    """
+    Default server has already been registered
+    """
+
+
+class SettingNotDefined(ConfigurationError):
+    """
+    Setting '%s' is not defined
+    """
+
+
+class NoProtocolsRegistered(ConfigurationError):
+    """
+    Server has no protocols
+    """
+
+
+class BadProtocol(ConfigurationError):
+    """
+    Bad protocol: %s
+    """
+
+
+class FunctionAlreadyRegistered(ConfigurationError):
+    """
+    Function '%s' is already registered
+    """
+
+
+class BadFunction(ConfigurationError):
+    """
+    Bad function: %s
+    """
 
 
 class ProtocolError(Error):
@@ -57,25 +72,23 @@ class ProtocolError(Error):
 
 
 class ProtocolNotFound(ClientErrorTypeMixin, ProtocolError):
-
-    def __str__(self):
-        return 'A protocol cannot be found: %s' % self.args[0]
+    """
+    Protocol '%s' cannot be found
+    """
 
 
 class BadRequest(ClientErrorTypeMixin, ProtocolError):
-
-    def __str__(self):
+    def __unicode__(self):
         if not self.args:
-            return 'Bad request'
-        return 'Bad request: %s' % self.args[0]
+            return u'Bad request'
+        return u'Bad request: %s' % self.args[0]
 
 
 class AccessDenied(ClientErrorTypeMixin, ProtocolError):
-
-    def __str__(self):
+    def __unicode__(self):
         if not self.args or not self.args[0]:
-            return 'Access denied'
-        return 'Access denied for user %s' % self.args[0]
+            return u'Access denied'
+        return u'Access denied for user %s' % self.args[0]
 
 
 class FunctionError(Error):
@@ -83,9 +96,9 @@ class FunctionError(Error):
 
 
 class FunctionNotFound(ClientErrorTypeMixin, FunctionError):
-
-    def __str__(self):
-        return 'A function cannot be found: %s' % self.args[0]
+    """
+    Function '%s' cannot be found
+    """
 
 
 class FieldError(ClientErrorTypeMixin, Error):
@@ -93,12 +106,12 @@ class FieldError(ClientErrorTypeMixin, Error):
 
 
 class MissingFieldValue(FieldError):
-
-    def __str__(self):
-        return 'The value of \'%s\' field is missing.' % self.args[0]
+    """
+    The value of field '%s' is missing.
+    """
 
 
 class WrongFieldValueType(FieldError):
-
-    def __str__(self):
-        return 'The value of \'%s\' field is of wrong type.' % self.args[0]
+    """
+    The value of field '%s' is of wrong type.
+    """