+++ src/pyws/utils.py
@@ -1,5 +1,18 @@
 from functools import wraps
 
+try:
+    from cgi import parse_qs
+except ImportError:
+    from urlparse import parse_qs
+
+try:
+    import json
+except ImportError:
+    import simplejson as json
+
+
+ENCODING = 'utf-8'
+
 
 def cache_method_result(attr):
     def decorator(func):
@@ -18,3 +31,9 @@ def cache_method_result(attr):
 
 def cached_property(func):
     return property(cache_method_result(func))
+
+
+class DefaultStrImplemntationMixin(object):
+
+    def __str__(self):
+        return unicode(self).encode('utf-8')