+++ tests/java/axis/testcases/RaisesExceptionTestCase.java
@@ -10,7 +10,7 @@ public class RaisesExceptionTestCase extends TestServiceTestCase {
     @Test
     public void test() {
         try {
-            System.out.println(port.raises_exception());
+            System.out.println(port.raises_exception("hello"));
         } catch (com.example.types.Error e) {
             Assert.assertTrue(e.toString().equals("hello error"));
             return;